Implemented 'Caching'.
The default 'cache' option is normally just passed to the $.ajax command to help prevent browsers from caching GET requests, etc.
That doesn't help us cache requests in to prevent multiple lookups.

I implemented a true 'per term' cache. So each term you enter is cached. Each cache can have a max age limit. By default I left it at 1ms, because the 'cache' option defaults to true we don't want to change the known behavior for existing users. Simply change the `cache_timeout` option to something more reasonable to make use of true caching.

I also tweaked the way options are passed into the `Select2EntityType` to make it easier to expand on this in the future (passing in 6 arguments to the constructor starts to get a little crazy).